refactor: moved CAN message sending functionality in separate module
